requirements.txt
wheel
pandas
numpy
ipykernel
notebook
rdkit
PyTDC
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
torch_geometric==2.3.1
h5torch
transformers==4.43.4
sentencepiece==0.2.0
fuse-med-ml==0.3.0
pytorch-fast-transformers==0.4.0